Fix notification serverless functions and refactorings. (#560)
* move metadata parsers to utils to share with serverless.
* fix GCP serverless
* Add note for ensuresResources.
* Update lambda case.
