+++ notification/src/api/notification/notification.controller.js
@@ -4,18 +4,9 @@ const httpUtils = require('../../utils/commons')
 const {
   processNotification,
 } = require('../../handler/notification/notification.handler')
-const config = require('../../config/config')
-const logger = require('../../utils/logger').getLogger()
+const { getCtpProjectConfig, getAdyenConfig } = require('../../utils/parser')
 
-class ValidationError extends Error {
-  constructor({ stack, message, notification, isRecoverable }) {
-    super()
-    this.stack = stack
-    this.message = message
-    this.notification = JSON.stringify(notification)
-    this.isRecoverable = isRecoverable
-  }
-}
+const logger = require('../../utils/logger').getLogger()
 
 async function handleNotification(request, response) {
   if (request.method !== 'POST') return httpUtils.sendResponse(response)
@@ -44,54 +35,6 @@ async function handleNotification(request, response) {
     }
     return sendAcceptedResponse(response)
   }
-}
-
-function getCtpProjectConfig(notification) {
-  let commercetoolsProjectKey
-  try {
-    commercetoolsProjectKey =
-      notification.NotificationRequestItem.additionalData[
-        'metadata.commercetoolsProjectKey'
-      ]
-  } catch (e) {
-    throw new ValidationError({
-      stack: e.stack,
-      notification,
-      message:
-        'Notification does not contain the field `metadata.commercetoolsProjectKey`.',
-      isRecoverable: false,
-    })
-  }
-
-  let ctpProjectConfig
-  try {
-    ctpProjectConfig = config.getCtpConfig(commercetoolsProjectKey)
-  } catch (e) {
-    throw new ValidationError({
-      stack: e.stack,
-      message: e.message,
-      notification,
-      isRecoverable: true,
-    })
-  }
-  return ctpProjectConfig
-}
-
-function getAdyenConfig(notification) {
-  const adyenMerchantAccount =
-    notification.NotificationRequestItem.merchantAccountCode
-  let adyenConfig
-  try {
-    adyenConfig = config.getAdyenConfig(adyenMerchantAccount)
-  } catch (e) {
-    throw new ValidationError({
-      stack: e.stack,
-      message: e.message,
-      notification,
-      isRecoverable: true,
-    })
-  }
-  return adyenConfig
 }
 
 function sendAcceptedResponse(response) {
