+++ notification/src/lambda.js
@@ -1,13 +1,9 @@
 const handler = require('./handler/notification/notification.handler')
-const config = require('./config/config')
 const logger = require('./utils/logger').getLogger()
-const setup = require('./config/init/ensure-interface-interaction-custom-type')
-
 const { getNotificationForTracking } = require('./utils/commons')
+const { getCtpProjectConfig, getAdyenConfig } = require('./utils/parser')
 
-let initialised = false
-
-exports.handler = async function (event) {
+exports.handler = async (event) => {
   const { notificationItems } = event
   if (!notificationItems) {
     const error = new Error('No notification received.')
@@ -20,37 +16,27 @@ exports.handler = async function (event) {
     )
     throw error
   }
-  for (const notification of notificationItems) {
-    try {
-      const commercetoolsProjectKey =
-        notification.NotificationRequestItem.additionalData[
-          'metadata.commercetoolsProjectKey'
-        ]
-      const adyenMerchantAccount =
-        notification.NotificationRequestItem.merchantAccountCode
-      const ctpProjectConfig = config.getCtpConfig(commercetoolsProjectKey)
-      const adyenConfig = config.getAdyenConfig(adyenMerchantAccount)
+  try {
+    for (const notification of notificationItems) {
+      const ctpProjectConfig = getCtpProjectConfig(notification)
+      const adyenConfig = getAdyenConfig(notification)
 
-      if (!initialised) {
-        await setup.ensureInterfaceInteractionCustomType(ctpProjectConfig)
-        initialised = true
-      }
       await handler.processNotification(
         notification,
         adyenConfig.enableHmacSignature,
         ctpProjectConfig
       )
-    } catch (e) {
-      logger.error(
-        {
-          notification: getNotificationForTracking(notification),
-          err: e,
-        },
-        'Unexpected error when processing event'
-      )
-      throw e
+    }
+  } catch (err) {
+    logger.error(
+      { notification: getNotificationForTracking(notificationItems), err },
+      'Unexpected exception occurred.'
+    )
+    if (err.isRecoverable) {
+      throw err
     }
   }
+
   return {
     notificationResponse: '[accepted]',
   }
