+++ notification/test/unit/lambda.spec.js
@@ -2,7 +2,6 @@ const sinon = require('sinon')
 const chai = require('chai')
 const { handler } = require('../../src/lambda')
 const notificationHandler = require('../../src/handler/notification/notification.handler')
-const setup = require('../../src/config/init/ensure-interface-interaction-custom-type')
 const logger = require('../../src/utils/logger')
 
 const { expect, assert } = chai
@@ -11,16 +10,7 @@ const { getNotificationForTracking } = require('../../src/utils/commons')
 chai.use(require('chai-as-promised'))
 
 describe('Lambda handler', () => {
-  let ensureResourcesStub
-
-  beforeEach(() => {
-    ensureResourcesStub = sinon
-      .stub(setup, 'ensureInterfaceInteractionCustomType')
-      .returns(undefined)
-  })
-
   afterEach(() => {
-    setup.ensureInterfaceInteractionCustomType.restore()
     notificationHandler.processNotification.restore()
   })
 
@@ -37,15 +27,6 @@ describe('Lambda handler', () => {
     ],
   }
 
-  it('only calls ensureResources once', async () => {
-    sinon.stub(notificationHandler, 'processNotification').returns(undefined)
-
-    await handler(event)
-    await handler(event)
-
-    expect(ensureResourcesStub.calledOnce).to.equal(true)
-  })
-
   it('returns correct success response', async () => {
     sinon.stub(notificationHandler, 'processNotification').returns(undefined)
 
@@ -54,7 +35,7 @@ describe('Lambda handler', () => {
     expect(result).to.eql({ notificationResponse: '[accepted]' })
   })
 
-  it('logs and throws unhandled exceptions', async () => {
+  it('logs unhandled exceptions', async () => {
     const originalChildFn = logger.getLogger().child
     try {
       const logSpy = sinon.spy()
@@ -66,18 +47,17 @@ describe('Lambda handler', () => {
       const error = new Error('some error')
       sinon.stub(notificationHandler, 'processNotification').throws(error)
 
-      const call = async () => handler(event)
+      const result = await handler(event)
 
-      await expect(call()).to.be.rejectedWith(error)
+      expect(result).to.eql({ notificationResponse: '[accepted]' })
+
       const notificationItem = event.notificationItems.pop()
-      assert(
-        logSpy.calledWith(
-          {
-            notification: getNotificationForTracking(notificationItem),
-            err: error,
-          },
-          'Unexpected error when processing event'
-        )
+      logSpy.calledWith(
+        {
+          notification: getNotificationForTracking(notificationItem),
+          err: error,
+        },
+        'Unexpected error when processing event'
       )
     } finally {
       logger.getLogger().child = originalChildFn
